requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
astroid==3.3.11
attrs==25.4.0
backoff==1.11.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
coverage==7.11.0
cryptography==46.0.3
dill==0.4.0
distro==1.9.0
Django==5.2.8
django-extensions==4.1
djangorestframework==3.15.2
dotenv==0.9.9
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.41
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isort==6.1.0
jiter==0.11.0
MarkupSafe==3.0.3
mccabe==0.7.0
multidict==6.7.0
openai==2.2.0
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.9
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
PyNaCl==1.6.0
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
setuptools==78.1.1
smmap==5.0.2
sniffio==1.3.1
spotify==0.10.2
spotipy==2.25.1
sqlparse==0.5.3
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
